fastapi/requirements.txt
# ...new file...
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
pandas
prophet